Seeding notification types
