+++ Examonimy/ExamonimyWeb/Entities/NotificationType.cs
@@ -1,0 +1,17 @@
+﻿using System.ComponentModel.DataAnnotations;
+using ExamonimyWeb.Utilities;
+
+namespace ExamonimyWeb.Entities
+{
+    public class NotificationType
+    {
+        [Key]
+        public int Id { get; set; }
+
+        [Required]
+        public required Entity Entity { get; set; }
+
+        [Required]
+        public required Operation Operation { get; set; }
+    }
+}
