+++ Examonimy/ExamonimyWeb/Utilities/Entity.cs
@@ -1,0 +1,8 @@
+﻿namespace ExamonimyWeb.Utilities
+{
+    public enum Entity
+    {
+        ExamPaperReviewer,
+        ExamPaperComment
+    }
+}
