+++ Examonimy/ExamonimyWeb/Utilities/Operation.cs
@@ -1,0 +1,10 @@
+﻿namespace ExamonimyWeb.Utilities
+{
+    public enum Operation
+    {
+        AskForReviewForExamPaper,
+        CommentExamPaper,
+        ApproveExamPaper,
+        RejectExamPaper
+    }
+}
